+++ src/curvesimilarities/frechet.py
@@ -124,7 +124,7 @@ def decision_problem(P, Q, epsilon):
 def significant_events(
     P,
     Q,
-    param="arc-length",
+    param_type="arc-length",
     rel_tol=0.0,
     abs_tol=float(EPSILON),
     event_rel_tol=0.0,
@@ -142,7 +142,7 @@ def significant_events(
     Q : array_like
         A :math:`q` by :math:`n` array of :math:`q` vertices of a polyline in an
         :math:`n`-dimensional space.
-    param : {'arc-length', 'vertex'}
+    param_type : {'arc-length', 'vertex'}
         Parametrization of matching.
     rel_tol, abs_tol : double
         Relative and absolute tolerances for parametric search of the feasible distance.
@@ -233,14 +233,14 @@ def significant_events(
     events = events[:count]
     errors = errors[:count]
 
-    if param == "arc-length":
+    if param_type == "arc-length":
         events = np.stack(
             (
                 index2arclength(P, events[:, :, 0].copy()),
                 index2arclength(Q, events[:, :, 1].copy()),
             )
         ).transpose(1, 2, 0)
-    elif param == "vertex":
+    elif param_type == "vertex":
         pass
     else:
         raise ValueError("Unknown option for parametrization.")
@@ -252,7 +252,7 @@ def significant_events(
 def fd_matching(
     P,
     Q,
-    param="arc-length",
+    param_type="arc-length",
     rel_tol=0.0,
     abs_tol=float(EPSILON),
     event_rel_tol=0.0,
@@ -271,7 +271,7 @@ def fd_matching(
     Q : array_like
         A :math:`q` by :math:`n` array of :math:`q` vertices of a polyline in an
         :math:`n`-dimensional space.
-    param : {'arc-length', 'vertex'}
+    param_type : {'arc-length', 'vertex'}
         Parametrization of matching.
     rel_tol, abs_tol : double
         Relative and absolute tolerances for parametric search of the Fréchet distance.
@@ -320,14 +320,14 @@ def fd_matching(
         dist = max(eps, np.linalg.norm(P[0] - Q[0]), np.linalg.norm(P[-1] - Q[-1]))
     else:
         dist = max(np.linalg.norm(P[0] - Q[0]), np.linalg.norm(P[-1] - Q[-1]))
-    if param == "arc-length":
+    if param_type == "arc-length":
         matching = np.stack(
             (
                 index2arclength(P, matching[:, 0].copy()),
                 index2arclength(Q, matching[:, 1].copy()),
             )
         ).T
-    elif param == "vertex":
+    elif param_type == "vertex":
         pass
     else:
         raise ValueError("Unknown option for parametrization.")